requirements.txt
Pillow==9.1.0
protobuf==3.20.1
requests==2.27.1
streamlit==1.3.1
google-cloud-storage